backend/requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.52.0
dnspython==2.5.0
httpx==0.25.2
tldextract==5.1.1
aiohttp==3.9.1
cryptography==42.0.2
python-whois==0.9.5
certifi==2024.2.2
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
alembic==1.13.1
email-validator==2.1.0
slowapi==0.1.9
limits[redis]==3.10.1
